Added some screens, extedned Game, and moved input controller
