+++ src/main/java/classes/BoardScreen.java
@@ -1,0 +1,40 @@
+package classes;
+
+import com.badlogic.gdx.*;
+
+public class BoardScreen implements Screen {
+    @Override
+    public void show() {
+
+    }
+
+    @Override
+    public void render(float v) {
+
+    }
+
+    @Override
+    public void resize(int i, int i1) {
+
+    }
+
+    @Override
+    public void pause() {
+
+    }
+
+    @Override
+    public void resume() {
+
+    }
+
+    @Override
+    public void hide() {
+
+    }
+
+    @Override
+    public void dispose() {
+
+    }
+}
